+++ src/components/form/MultivalueField/multivalueField.js
@@ -58,6 +58,10 @@ export function addField(scope, initialValue = '') {
 }
 
 export function activateMultivalueField(scope) {
+  if (scope.hasAttribute('pl-listener-assigned')) {
+    return;
+  }
+
   scope.querySelector('.multivalue-field__add-entity').addEventListener('click', function (e) {
     e.preventDefault();
     addField(scope, '');
@@ -70,16 +74,29 @@ export function activateMultivalueField(scope) {
   // Set mapping between labels and inputs.
   const labels = scope.querySelectorAll('label');
   for (const label of labels) {
-    setMappingForLabel(label);
+    setLabelMappingForInput(label);
   }
+
+  scope.setAttribute('pl-listener-assigned', '1');
 }
 
-export function setMappingForLabel(label) {
+export function setLabelMappingForInput(label) {
   const input = label.closest('.input-field');
 
   if (input) {
     const id = generateHash();
     input.querySelector('input').setAttribute('id', id);
+    label.setAttribute('for', id);
+    label.setAttribute('id', `${id}-label`);
+  }
+}
+
+export function setLabelMappingForSelect(label) {
+  const select = label.closest('.dropdown');
+
+  if (select) {
+    const id = generateHash();
+    select.querySelector('select').setAttribute('id', id);
     label.setAttribute('for', id);
     label.setAttribute('id', `${id}-label`);
   }
@@ -104,7 +121,7 @@ function buildDeleteButton(scope) {
       dispatchMultigroupEvent(group);
     }
 
-    setMappingForLabel(label);
+    setLabelMappingForInput(label);
 
     scope.setAttribute('data-count', scope.getAttribute('data-count') - 1);
   });